Change user Id from int to long

+++ src/PasswordManager.Application/ApplicationSettings.cs
@@ -1,6 +1,6 @@
 ﻿
 namespace PasswordManager.Application {
 	public class ApplicationSettings {
-		public int[] AdminIds { get; set; }
+		public long[] AdminIds { get; set; }
 	}
 }
