+++ src/PasswordManager.Application/Services/Abstractions/IUserService.cs
@@ -6,36 +6,36 @@ namespace PasswordManager.Application.Services.Abstractions {
 
 	public interface IUserService {
 
-		Task<User> AddUserAsync(int userId, string langCode);
+		Task<User> AddUserAsync(long userId, string langCode);
 
 		/// <summary></summary>
 		/// <returns>User with Action and Lang</returns>
-		Task<User> GetUserActionAsync(int userId);
+		Task<User> GetUserActionAsync(long userId);
 
 		/// <summary></summary>
 		/// <returns>List of user Ids and theis Accounts Count</returns>
 		Task<IList<User>> GetAllBasicInfoAsync();
 
-		Task UpdateActionAsync(int userId, UserAction action);
+		Task UpdateActionAsync(long userId, UserAction action);
 		
 		Task UpdateActionAsync(User user);
 		
-		Task<string> GetKeyHint(int userId);
+		Task<string> GetKeyHint(long userId);
 		
-		Task UpdateKeyHint(int userId, string keyHint);
+		Task UpdateKeyHint(long userId, string keyHint);
 
-		Task<string> GetPasswordGeneratorPattern(int userId);
+		Task<string> GetPasswordGeneratorPattern(long userId);
 
-		Task UpdatePasswordGeneratorPattern(int userId, string passwordGeneratorPattern);
+		Task UpdatePasswordGeneratorPattern(long userId, string passwordGeneratorPattern);
 		
-		Task UpdateLanguage(int userId, string langCode);
+		Task UpdateLanguage(long userId, string langCode);
 		
 		/// <summary>
 		///     Deletes user if he is not admin
 		/// </summary>
-		Task<bool> DeleteUser(int userId);
+		Task<bool> DeleteUser(long userId);
 
-		Task<User> GetUserOutdatedTimeAsync(int userId);
+		Task<User> GetUserOutdatedTimeAsync(long userId);
 
 	}
