+++ src/PasswordManager.Bot/Services/PasswordDecryptionService.cs
@@ -4,19 +4,19 @@ using PasswordManager.Core.Entities;
 
 namespace PasswordManager.Bot.Services {
 	public class PasswordDecryptionService : IPasswordDecryptionService {
-		private readonly Dictionary<int, Account> accounts;
+		private readonly Dictionary<long, Account> accounts;
 
 		public PasswordDecryptionService() {
-			accounts = new Dictionary<int, Account>();
+			accounts = new Dictionary<long, Account>();
 		}
 
-		public void StartDecryptionRequest(int userId, Account account)
+		public void StartDecryptionRequest(long userId, Account account)
 			=> accounts[userId] = account;
 
-		public void FinishDecryptionRequest(int userId)
+		public void FinishDecryptionRequest(long userId)
 			=> accounts.Remove(userId);
 
-		public Account GetAccount(int userId) {
+		public Account GetAccount(long userId) {
 			if(accounts.TryGetValue(userId, out Account account))
 				return account;
 			return null;
